Add enemy death functionality
The enemy has a death collider set as a triger, placed on the top of itself. EnemyController checks for trigger collisions, and the collider has the tag 'Player', it calls its Kill method, which then calls its implemented OnDeath method. The OnDeath method destroys the object. We can implement the animation there.